searcher: separate out get_list_directory() and friends
closes #97
